Crate servo_constellation_traits

The interface to the Constellation, which prevents other crates from depending directly on the constellation crate itself. In addition to all messages to the Constellation, this crate is responsible for defining types that cross the process boundary from the embedding/rendering layer all the way to script, thus it should have very minimal dependencies on other parts of Servo.

Modules

from_script_message 🔒

Messages send from the ScriptThread to the Constellation.

structured_data 🔒

This module contains implementations of structured data as described in https://html.spec.whatwg.org/multipage/#safe-passing-of-structured-data

Structs

AuxiliaryWebViewCreationRequest

Specifies the information required to load an auxiliary browsing context.

AuxiliaryWebViewCreationResponse

Constellation’s response to auxiliary browsing context creation requests.

BlobImpl

The data backing a DOM Blob.

BroadcastChannelMsg

Message for communication between the constellation and a global managing broadcast channels.

DOMMessage

Message that gets passed to service worker scope on postMessage

DomException

A serializable version of the DOMException interface.

DomMatrix

A serializable version of the DOMMatrix/DOMMatrixReadOnly interface.

DomPoint

A serializable version of the DOMPoint/DOMPointReadOnly interface.

DomQuad

A serializable version of the DOMQuad interface.

DomRect

A serializable version of the DOMRect/DOMRectReadOnly interface.

FileBlob

File-based blob

IFrameLoadInfo

Specifies the information required to load an iframe.

IFrameLoadInfoWithData

Specifies the information required to load a URL in an iframe.

IFrameSizeMsg

An iframe sizing operation.

Job

https://w3c.github.io/ServiceWorker/#dfn-job

LoadData

can be passed to LoadUrl to load a page with GET/POST parameters or headers

MessagePortImpl

The data and logic backing the DOM managed MessagePort.

PortMessageTask

A task on the https://html.spec.whatwg.org/multipage/#port-message-queue

PortTransferInfo

The information needed by a global to process the transfer of a port.

SWManagerSenders

Channels to allow service worker manager to communicate with constellation and resource thread

ScopeThings

Entities required to spawn service workers

ScriptToConstellationChan

A Script to Constellation channel.

ScrollStateUpdate

A data structure which contains information for the pipeline after a scroll happens in the embedder-side WebView.

SerializableAesKeyAlgorithm

A serializable version of the AesKeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableAlgorithm

A serializable version of the Algorithm dictionary, used by the SubtleCrypto interface.

SerializableCShakeParams

A serializable version of the CShakeParams dictionary, used by the SubtleCrypto interface.

SerializableCryptoKey

A serializable version of the CryptoKey interface.

SerializableEcKeyAlgorithm

A serializable version of the EcKeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableFile

SerializableFileList

SerializableHmacKeyAlgorithm

A serializable version of the HmacKeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableImageBitmap

A serializable version of the ImageBitmap interface.

SerializableImageData

SerializableIter

An iterator over the variants of Serializable

SerializableKangarooTwelveParams

A serializable version of the KangarooTwelvelParams dictionary, used by the SubtleCrypto interface.

SerializableKeyAlgorithm

A serializable version of the KeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableKmacKeyAlgorithm

A serializable version of the KmacKeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableQuotaExceededError

A serializable version of the QuotaExceededError interface.

SerializableRsaHashedKeyAlgorithm

A serializable version of the RsaHashedKeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableTurboShakeParams

A serializable version of the TurboShakeParams dictionary, used by the SubtleCrypto interface.

ServiceWorkerRegistrationInfo

https://w3c.github.io/ServiceWorker/#dfn-service-worker-registration

SessionHistoryTraversalRequest

A history traversal request.

StructuredSerializedData

A data-holder for serialized data and transferred objects. https://html.spec.whatwg.org/multipage/#structuredserializewithtransfer

TargetSnapshotParams

https://html.spec.whatwg.org/multipage/#target-snapshot-params

TransferableOffscreenCanvas

A struct supporting the transfer of OffscreenCanvas, which loosely corresponds to the dataHolder in https://html.spec.whatwg.org/multipage/#the-offscreencanvas-interface:offscreencanvas-16

TransferrableIter

An iterator over the variants of Transferrable

TransformStreamData

WorkerAnimationFrameTick

Message delivered to a worker event loop to run animation frame callbacks.

WorkerGlobalScopeInit

Resources required by workerglobalscopes

WorkerScriptLoadOrigin

Common entities representing a network load origin

Enums

BlobData

Different backends of Blob

ConstellationInterest

Identifies a category of events/notifications that a pipeline can register interest in with the constellation. When a pipeline has active listeners for events in a given category, it registers interest so the constellation only sends notifications to pipelines that care.

EmbedderToConstellationMessage

Messages to the Constellation from the embedding layer, whether from ServoRenderer or from libservo itself.

HistoryTraversalSource

The source of a [HistoryTraversalRequest].

JobError

The kind of error the job promise should be rejected with.

JobResult

Messages sent from Job algorithms steps running in the SW manager, in order to resolve or reject the job promise.

JobResultValue

Jobs are resolved with the help of various values.

JobType

https://w3c.github.io/ServiceWorker/#dfn-job-type

KeyboardScroll

An enum that describe a type of keyboard scroll.

LoadOrigin

The origin where a given load was initiated. Useful for origin checks, for example before evaluation a JS URL.

LogEntry

A log entry reported to the constellation We don’t report all log entries, just serious ones. We need a separate type for this because LogLevel isn’t serializable.

MessagePortMsg

Messages for communication between the constellation and a global managing ports.

NavigationHistoryBehavior

https://html.spec.whatwg.org/multipage/#navigation-supporting-concepts:navigationhistorybehavior

PaintMetricEvent

A description of a paint metric that is sent from the Servo renderer to the constellation.

RemoteFocusOperation

The type of focus operation to do on a remote document.

ScreenshotReadinessResponse

ScriptToConstellationMessage

Messages from the script to the constellation.

SequentialFocusDirection

https://html.spec.whatwg.org/multipage/#sequential-focus-direction

Serializable

All the DOM interfaces that can be serialized.

SerializableCryptoKeyHandle

A serializable version of the Handle type, used by the CryptoKey interface.

SerializableDigestAlgorithm

A serializable version of the DigestAlgorithm type, used the SubtleCrypto interface.

SerializableKeyAlgorithmAndDerivatives

A serializable version of the KeyAlgorithmAndDerivatives type, used by the SubtleCrypto interface.

ServiceWorkerAlgorithm

https://w3c.github.io/ServiceWorker/#algorithms

ServiceWorkerAlgorithmResult

https://w3c.github.io/ServiceWorker/#algorithms

ServiceWorkerMsg

Messages sent to Service Worker Manager thread

Transferrable

All the DOM interfaces that can be transferred.

TraversalDirection

The direction of a history traversal

UserContentManagerAction

WindowSizeType

The type of window size change.

Traits

ServiceWorkerManagerFactory

This trait allows creating a ServiceWorkerManager without depending on the script crate.

Type Aliases

ScriptToConstellationSender