servo_constellation_traits

Enum EmbedderToConstellationMessage

pub enum EmbedderToConstellationMessage {

    Exit,
    AllowNavigationResponse(PipelineId, bool),
    LoadUrl(WebViewId, UrlRequest),
    TraverseHistory(SessionHistoryTraversalRequest),
    ChangeViewportDetails(WebViewId, ViewportDetails, WindowSizeType),
    ThemeChange(WebViewId, Theme),
    TickAnimation(Vec<WebViewId>),
    NoLongerWaitingOnAsynchronousImageUpdates(Vec<PipelineId>),
    WebDriverCommand(WebDriverCommandMsg),
    Reload(WebViewId),
    LogEntry(Option<ScriptEventLoopId>, Option<String>, LogEntry),
    NewWebView(ServoUrl, NewWebViewDetails),
    CloseWebView(WebViewId),
    FocusWebView(WebViewId),
    BlurWebView,
    ForwardInputEvent(WebViewId, InputEventAndId, Option<PaintHitTestResult>),
    RefreshCursor(PipelineId),
    ExitFullScreen(WebViewId),
    MediaSessionAction(MediaSessionActionType),
    SetWebViewThrottled(WebViewId, bool),
    SetScrollStates(PipelineId, ScrollStateUpdate),
    PaintMetric(PipelineId, PaintMetricEvent),
    EvaluateJavaScript(WebViewId, JavaScriptEvaluationId, String),
    CreateMemoryReport(GenericCallback<MemoryReportResult>),
    SendImageKeysForPipeline(PipelineId, Vec<ImageKey>),
    PreferencesUpdated(Vec<(&'static str, PrefValue)>),
    RequestScreenshotReadiness(WebViewId),
    EmbedderControlResponse(EmbedderControlId, EmbedderControlResponse),
    UserContentManagerAction(UserContentManagerId, UserContentManagerAction),
    UpdatePinchZoomInfos(PipelineId, PinchZoomInfos),
    SetAccessibilityActive(WebViewId, bool),

}

Messages to the Constellation from the embedding layer, whether from ServoRenderer or from libservo itself.

Variants

Exit

Exit the constellation.

AllowNavigationResponse(PipelineId, bool)

Whether to allow script to navigate.

LoadUrl(WebViewId, UrlRequest)

Request to load a page, with optionally additional data in [URLRequest].

TraverseHistory(SessionHistoryTraversalRequest)

Request to traverse the joint session history of the provided browsing context.

ChangeViewportDetails(WebViewId, ViewportDetails, WindowSizeType)

Inform the Constellation that a WebView’s ViewportDetails have changed.

ThemeChange(WebViewId, Theme)

Inform the constellation of a theme change.

TickAnimation(Vec<WebViewId>)

Requests that the constellation instruct script/layout to try to layout again and tick animations.

NoLongerWaitingOnAsynchronousImageUpdates(Vec<PipelineId>)

Notify the ScriptThread that the Servo renderer is no longer waiting on asynchronous image uploads for the given Pipeline. These are mainly used by canvas to perform uploads while the display list is being built.

WebDriverCommand(WebDriverCommandMsg)

Dispatch a webdriver command

Reload(WebViewId)

Reload a top-level browsing context.

LogEntry(Option<ScriptEventLoopId>, Option<String>, LogEntry)

A log entry, with the top-level browsing context id and thread name

NewWebView(ServoUrl, NewWebViewDetails)

Create a new top level browsing context.

CloseWebView(WebViewId)

Close a top level browsing context.

FocusWebView(WebViewId)

Make a webview focused. [EmbedderMsg::WebViewFocused] will be sent with the result of this operation.

BlurWebView

Make none of the webviews focused.

ForwardInputEvent(WebViewId, InputEventAndId, Option<PaintHitTestResult>)

Forward an input event to an appropriate ScriptTask.

RefreshCursor(PipelineId)

Request that the given pipeline refresh the cursor by doing a hit test at the most recently hovered cursor position and resetting the cursor. This happens after a display list update is rendered.

ExitFullScreen(WebViewId)

Request to exit from fullscreen mode

MediaSessionAction(MediaSessionActionType)

Media session action.

SetWebViewThrottled(WebViewId, bool)

Set whether to use less resources, by stopping animations and running timers at a heavily limited rate.

SetScrollStates(PipelineId, ScrollStateUpdate)

The Servo renderer scrolled and is updating the scroll states of the nodes in the given pipeline via the constellation.

PaintMetric(PipelineId, PaintMetricEvent)

Notify the constellation that a particular paint metric event has happened for the given pipeline.

EvaluateJavaScript(WebViewId, JavaScriptEvaluationId, String)

Evaluate a JavaScript string in the context of a WebView. When execution is complete or an error is encountered, a correpsonding message will be sent to the embedding layer.

CreateMemoryReport(GenericCallback<MemoryReportResult>)

Create a memory report and return it via the GenericCallback

SendImageKeysForPipeline(PipelineId, Vec<ImageKey>)

Sends the generated image key to the image cache associated with this pipeline.

PreferencesUpdated(Vec<(&'static str, PrefValue)>)

A set of preferences were updated with the given new values.

RequestScreenshotReadiness(WebViewId)

Request preparation for a screenshot of the given WebView. The Constellation will send a message to the Embedder when the screenshot is ready to be taken.

EmbedderControlResponse(EmbedderControlId, EmbedderControlResponse)

A response to a request to show an embedder user interface control.

UserContentManagerAction(UserContentManagerId, UserContentManagerAction)

An action to perform on the given UserContentManagerId.

UpdatePinchZoomInfos(PipelineId, PinchZoomInfos)

Update pinch zoom details stored in the top level window

SetAccessibilityActive(WebViewId, bool)

Activate or deactivate accessibility features for the given WebView.