servo_constellation_traits

Module from_script_message

Messages send from the ScriptThread to the Constellation.

Structs

AuxiliaryWebViewCreationRequest

Specifies the information required to load an auxiliary browsing context.

AuxiliaryWebViewCreationResponse

Constellation’s response to auxiliary browsing context creation requests.

DOMMessage

Message that gets passed to service worker scope on postMessage

IFrameLoadInfo

Specifies the information required to load an iframe.

IFrameLoadInfoWithData

Specifies the information required to load a URL in an iframe.

IFrameSizeMsg

An iframe sizing operation.

Job

https://w3c.github.io/ServiceWorker/#dfn-job

LoadData

can be passed to LoadUrl to load a page with GET/POST parameters or headers

SWManagerSenders

Channels to allow service worker manager to communicate with constellation and resource thread

ScopeThings

Entities required to spawn service workers

ScriptToConstellationChan

A Script to Constellation channel.

ServiceWorkerRegistrationInfo

https://w3c.github.io/ServiceWorker/#dfn-service-worker-registration

TargetSnapshotParams

https://html.spec.whatwg.org/multipage/#target-snapshot-params

WorkerAnimationFrameTick

Message delivered to a worker event loop to run animation frame callbacks.

WorkerGlobalScopeInit

Resources required by workerglobalscopes

WorkerScriptLoadOrigin

Common entities representing a network load origin

Enums

ConstellationInterest

Identifies a category of events/notifications that a pipeline can register interest in with the constellation. When a pipeline has active listeners for events in a given category, it registers interest so the constellation only sends notifications to pipelines that care.

JobError

The kind of error the job promise should be rejected with.

JobResult

Messages sent from Job algorithms steps running in the SW manager, in order to resolve or reject the job promise.

JobResultValue

Jobs are resolved with the help of various values.

JobType

https://w3c.github.io/ServiceWorker/#dfn-job-type

KeyboardScroll

An enum that describe a type of keyboard scroll.

LoadOrigin

The origin where a given load was initiated. Useful for origin checks, for example before evaluation a JS URL.

NavigationHistoryBehavior

https://html.spec.whatwg.org/multipage/#navigation-supporting-concepts:navigationhistorybehavior

RemoteFocusOperation

The type of focus operation to do on a remote document.

ScreenshotReadinessResponse

ScriptToConstellationMessage

Messages from the script to the constellation.

SequentialFocusDirection

https://html.spec.whatwg.org/multipage/#sequential-focus-direction

ServiceWorkerAlgorithm

https://w3c.github.io/ServiceWorker/#algorithms

ServiceWorkerAlgorithmResult

https://w3c.github.io/ServiceWorker/#algorithms

ServiceWorkerMsg

Messages sent to Service Worker Manager thread

Traits

ServiceWorkerManagerFactory

This trait allows creating a ServiceWorkerManager without depending on the script crate.

Type Aliases

ScriptToConstellationSender