servo_constellation_traits::from_script_message

Enum ScriptToConstellationMessage

pub enum ScriptToConstellationMessage {

    ServiceWorkerAlgorithm(ServiceWorkerAlgorithm),
    CompleteMessagePortTransfer(MessagePortRouterId, Vec<MessagePortId>),
    MessagePortTransferResult(Option<MessagePortRouterId>, Vec<MessagePortId>, FxHashMap<MessagePortId, PortTransferInfo>),
    NewMessagePort(MessagePortRouterId, MessagePortId),
    NewMessagePortRouter(MessagePortRouterId, GenericCallback<MessagePortMsg>),
    RemoveMessagePortRouter(MessagePortRouterId),
    RerouteMessagePort(MessagePortId, PortMessageTask),
    MessagePortShipped(MessagePortId),
    EntanglePorts(MessagePortId, MessagePortId),
    DisentanglePorts(MessagePortId, Option<MessagePortId>),
    NewBroadcastChannelRouter(BroadcastChannelRouterId, GenericCallback<BroadcastChannelMsg>, ImmutableOrigin),
    RemoveBroadcastChannelRouter(BroadcastChannelRouterId, ImmutableOrigin),
    NewBroadcastChannelNameInRouter(BroadcastChannelRouterId, String, ImmutableOrigin),
    RemoveBroadcastChannelNameInRouter(BroadcastChannelRouterId, String, ImmutableOrigin),
    ScheduleBroadcast(BroadcastChannelRouterId, BroadcastChannelMsg),
    RegisterInterest(ConstellationInterest),
    UnregisterInterest(ConstellationInterest),
    BroadcastStorageEvent(WebStorageType, ServoUrl, Option<String>, Option<String>, Option<String>),
    ChangeRunningAnimationsState(AnimationState),
    RegisterWorkerAnimationFrameProvider(WorkerId, GenericSender<WorkerAnimationFrameTick>),
    UnregisterWorkerAnimationFrameProvider(WorkerId),
    ChangeWorkerAnimationFrameProviderState(WorkerId, bool),
    CreateCanvasPaintThread(Size2D<u64>, GenericSender<Option<(GenericSender<CanvasMsg>, CanvasId)>>),
    FocusAncestorBrowsingContextsForFocusingSteps(Option<BrowsingContextId>, FocusSequenceNumber),
    FocusRemoteBrowsingContext(BrowsingContextId, RemoteFocusOperation),
    GetTopForBrowsingContext(BrowsingContextId, GenericSender<Option<WebViewId>>),
    GetBrowsingContextInfo(PipelineId, GenericSender<Option<(BrowsingContextId, Option<PipelineId>)>>),
    GetChildBrowsingContextId(BrowsingContextId, usize, GenericSender<Option<BrowsingContextId>>),
    GetDocumentOrigin(PipelineId, GenericSender<Option<OriginSnapshot>>),
    IsCurrentlyFullyActive(PipelineId, GenericSender<bool>),
    GetInternalAncestorOriginObjectsList(PipelineId, GenericSender<Option<Vec<ImmutableOrigin>>>),
    LoadComplete,
    LoadUrl(LoadData, NavigationHistoryBehavior, TargetSnapshotParams),
    AbortLoadUrl,
    PostMessage {
        target: BrowsingContextId,
        source: PipelineId,
        target_origin: Option<ImmutableOrigin>,
        source_origin: ImmutableOrigin,
        data: StructuredSerializedData,
    },
    NavigatedToFragment(ServoUrl, NavigationHistoryBehavior),
    TraverseHistory(SessionHistoryTraversalRequest),
    PushHistoryState(HistoryStateId, ServoUrl),
    ReplaceHistoryState(HistoryStateId, ServoUrl),
    JointSessionHistoryLength(GenericSender<u32>),
    RemoveIFrame(BrowsingContextId, GenericSender<Vec<PipelineId>>),
    SetThrottledComplete(bool),
    ScriptLoadedURLInIFrame(IFrameLoadInfoWithData),
    ScriptNewIFrame(IFrameLoadInfoWithData),
    CreateAuxiliaryWebView(AuxiliaryWebViewCreationRequest),
    ActivateDocument,
    SetFinalUrl(ServoUrl),
    LogEntry(Option<ScriptEventLoopId>, Option<String>, LogEntry),
    DiscardDocument,
    DiscardTopLevelBrowsingContext,
    PipelineExited,
    ForwardDOMMessage(DOMMessage, ServoUrl),
    MediaSessionEvent(PipelineId, MediaSessionEvent),
    RequestAdapter(GenericCallback<WebGPUAdapterResponse>, RequestAdapterOptions, AdapterId),
    GetWebGPUChan(GenericSender<Option<WebGPU>>),
    TitleChanged(PipelineId, String),
    IFrameSizes(Vec<IFrameSizeMsg>),
    ReportMemory(GenericCallback<MemoryReportResult>),
    FinishJavaScriptEvaluation(JavaScriptEvaluationId, Result<JSValue, JavaScriptEvaluationError>),
    ForwardKeyboardScroll(PipelineId, KeyboardScroll),
    RespondToScreenshotReadinessRequest(ScreenshotReadinessResponse),
    TriggerGarbageCollection,
    AcquireWakeLock(WakeLockType),
    ReleaseWakeLock(WakeLockType),

}

Messages from the script to the constellation.

Variants

ServiceWorkerAlgorithm(ServiceWorkerAlgorithm)

CompleteMessagePortTransfer(MessagePortRouterId, Vec<MessagePortId>)

Request to complete the transfer of a set of ports to a router.

MessagePortTransferResult(Option<MessagePortRouterId>, Vec<MessagePortId>, FxHashMap<MessagePortId, PortTransferInfo>)

The results of attempting to complete the transfer of a batch of ports.

NewMessagePort(MessagePortRouterId, MessagePortId)

A new message-port was created or transferred, with corresponding control-sender.

NewMessagePortRouter(MessagePortRouterId, GenericCallback<MessagePortMsg>)

A global has started managing message-ports

RemoveMessagePortRouter(MessagePortRouterId)

A global has stopped managing message-ports

RerouteMessagePort(MessagePortId, PortMessageTask)

A task requires re-routing to an already shipped message-port.

MessagePortShipped(MessagePortId)

A message-port was shipped, let the entangled port know.

EntanglePorts(MessagePortId, MessagePortId)

Entangle two message-ports.

DisentanglePorts(MessagePortId, Option<MessagePortId>)

Disentangle two message-ports. The first is the initiator, the second the other port, unless the message is sent to complete a disentanglement, in which case the first one is the other port, and the second is none.

NewBroadcastChannelRouter(BroadcastChannelRouterId, GenericCallback<BroadcastChannelMsg>, ImmutableOrigin)

A global has started managing broadcast-channels.

RemoveBroadcastChannelRouter(BroadcastChannelRouterId, ImmutableOrigin)

A global has stopped managing broadcast-channels.

NewBroadcastChannelNameInRouter(BroadcastChannelRouterId, String, ImmutableOrigin)

A global started managing broadcast channels for a given channel-name.

RemoveBroadcastChannelNameInRouter(BroadcastChannelRouterId, String, ImmutableOrigin)

A global stopped managing broadcast channels for a given channel-name.

ScheduleBroadcast(BroadcastChannelRouterId, BroadcastChannelMsg)

Broadcast a message to all same-origin broadcast channels, excluding the source of the broadcast.

RegisterInterest(ConstellationInterest)

Register this pipeline’s interest in a category of notifications. The constellation will only send notifications in this category to pipelines that have registered interest.

UnregisterInterest(ConstellationInterest)

Unregister this pipeline’s interest in a category of notifications.

BroadcastStorageEvent(WebStorageType, ServoUrl, Option<String>, Option<String>, Option<String>)

Broadcast a storage event to every same-origin pipeline. The strings are key, old value and new value.

ChangeRunningAnimationsState(AnimationState)

Indicates whether this pipeline is currently running animations.

RegisterWorkerAnimationFrameProvider(WorkerId, GenericSender<WorkerAnimationFrameTick>)

Register a dedicated worker that can receive animation frame ticks.

UnregisterWorkerAnimationFrameProvider(WorkerId)

Unregister a dedicated worker animation frame provider.

ChangeWorkerAnimationFrameProviderState(WorkerId, bool)

Indicates whether a dedicated worker has pending animation frame callbacks.

CreateCanvasPaintThread(Size2D<u64>, GenericSender<Option<(GenericSender<CanvasMsg>, CanvasId)>>)

Requests that a new 2D canvas thread be created. (This is done in the constellation because 2D canvases may use the GPU and we don’t want to give untrusted content access to the GPU.)

FocusAncestorBrowsingContextsForFocusingSteps(Option<BrowsingContextId>, FocusSequenceNumber)

Notifies the constellation that this pipeline is requesting focus.

When this message is sent, the sender pipeline has already its local focus state updated. The constellation, after receiving this message, will broadcast messages to other pipelines that are affected by this focus operation.

The first field contains the browsing context ID of the container element if one was focused.

The second field is a sequence number that the constellation should use when sending a focus-related message to the sender pipeline next time.

FocusRemoteBrowsingContext(BrowsingContextId, RemoteFocusOperation)

Focus a remote BrowsingContext and run the focusing steps. This is used in two situations:

When calling the DOM focus() API on a remote Window as well as from WebDriver. The difference between this and FocusDocumentAsPartOfFocusingSteps is that this version actually does run the focusing steps and may result in blur and focus events firing up the frame tree.

When doing sequential focus navigation into and out of frames.

GetTopForBrowsingContext(BrowsingContextId, GenericSender<Option<WebViewId>>)

Get the top-level browsing context info for a given browsing context.

GetBrowsingContextInfo(PipelineId, GenericSender<Option<(BrowsingContextId, Option<PipelineId>)>>)

Get the browsing context id of the browsing context in which pipeline is embedded and the parent pipeline id of that browsing context.

GetChildBrowsingContextId(BrowsingContextId, usize, GenericSender<Option<BrowsingContextId>>)

Get the nth child browsing context ID for a given browsing context, sorted in tree order.

GetDocumentOrigin(PipelineId, GenericSender<Option<OriginSnapshot>>)

Get the origin of the document corresponding to the given pipeline

IsCurrentlyFullyActive(PipelineId, GenericSender<bool>)

If the document corresponding to the given pipeline is fully active

GetInternalAncestorOriginObjectsList(PipelineId, GenericSender<Option<Vec<ImmutableOrigin>>>)

Get the internal ancestor origin objects list of the document corresponding to the given pipeline

LoadComplete

All pending loads are complete, and the load event for this pipeline has been dispatched.

LoadUrl(LoadData, NavigationHistoryBehavior, TargetSnapshotParams)

A new load has been requested, with an option to replace the current entry once loaded instead of adding a new entry.

AbortLoadUrl

Abort loading after sending a LoadUrl message.

PostMessage

Post a message to the currently active window of a given browsing context.

Fields

target: BrowsingContextId

The target of the posted message.

source: PipelineId

The source of the posted message.

target_origin: Option<ImmutableOrigin>

The expected origin of the target.

source_origin: ImmutableOrigin

The source origin of the message. https://html.spec.whatwg.org/multipage/#dom-messageevent-origin

data: StructuredSerializedData

The data to be posted.

NavigatedToFragment(ServoUrl, NavigationHistoryBehavior)

Inform the constellation that a fragment was navigated to and whether or not it was a replacement navigation.

TraverseHistory(SessionHistoryTraversalRequest)

HTMLIFrameElement Forward or Back traversal.

PushHistoryState(HistoryStateId, ServoUrl)

Inform the constellation of a pushed history state.

ReplaceHistoryState(HistoryStateId, ServoUrl)

Inform the constellation of a replaced history state.

JointSessionHistoryLength(GenericSender<u32>)

Gets the length of the joint session history from the constellation.

RemoveIFrame(BrowsingContextId, GenericSender<Vec<PipelineId>>)

Notification that this iframe should be removed. Returns a list of pipelines which were closed.

SetThrottledComplete(bool)

Successful response to [crate::ConstellationControlMsg::SetThrottled].

ScriptLoadedURLInIFrame(IFrameLoadInfoWithData)

A load has been requested in an IFrame.

ScriptNewIFrame(IFrameLoadInfoWithData)

A load of the initial about:blank has been completed in an IFrame.

CreateAuxiliaryWebView(AuxiliaryWebViewCreationRequest)

Script has opened a new auxiliary browsing context.

ActivateDocument

Mark a new document as active

SetFinalUrl(ServoUrl)

Update the pipeline Url, which can change after redirections.

LogEntry(Option<ScriptEventLoopId>, Option<String>, LogEntry)

A log entry, with the top-level browsing context id and thread name

DiscardDocument

Discard the document.

DiscardTopLevelBrowsingContext

Discard the browsing context.

PipelineExited

Notifies the constellation that this pipeline has exited.

ForwardDOMMessage(DOMMessage, ServoUrl)

Send messages from postMessage calls from serviceworker to constellation for storing in service worker manager

MediaSessionEvent(PipelineId, MediaSessionEvent)

Notifies the constellation about media session events (i.e. when there is metadata for the active media session, playback state changes…).

RequestAdapter(GenericCallback<WebGPUAdapterResponse>, RequestAdapterOptions, AdapterId)

Create a WebGPU Adapter instance

GetWebGPUChan(GenericSender<Option<WebGPU>>)

Get WebGPU channel

TitleChanged(PipelineId, String)

Notify the constellation of a pipeline’s document’s title.

IFrameSizes(Vec<IFrameSizeMsg>)

Notify the constellation that the size of some <iframe>s has changed.

ReportMemory(GenericCallback<MemoryReportResult>)

Request results from the memory reporter.

FinishJavaScriptEvaluation(JavaScriptEvaluationId, Result<JSValue, JavaScriptEvaluationError>)

Return the result of the evaluated JavaScript with the given JavaScriptEvaluationId.

ForwardKeyboardScroll(PipelineId, KeyboardScroll)

Forward a keyboard scroll operation from an <iframe> to a parent pipeline.

RespondToScreenshotReadinessRequest(ScreenshotReadinessResponse)

Notify the Constellation of the screenshot readiness of a given pipeline.

TriggerGarbageCollection

Request the constellation to force garbage collection in all ScriptThread’s.

AcquireWakeLock(WakeLockType)

Request to acquire a wake lock of the given type. The constellation will track the aggregate lock count and notify the provider only when the count transitions from 0 to 1. https://w3c.github.io/screen-wake-lock/#dfn-acquire-wake-lock

ReleaseWakeLock(WakeLockType)

Request to release a wake lock of the given type. The constellation will track the aggregate lock count and notify the provider only when the count transitions from N to 0. https://w3c.github.io/screen-wake-lock/#dfn-release-wake-lock