servo_constellation_traits

Module structured_data

This module contains implementations of structured data as described in https://html.spec.whatwg.org/multipage/#safe-passing-of-structured-data

Re-exports

pub use serializable::*;

pub use transferable::*;

Modules

serializable 🔒

This module contains implementations in script that are serializable, as per https://html.spec.whatwg.org/multipage/#serializable-objects. The implementations are here instead of in script as they need to be passed through the Constellation.

transferable 🔒

This module contains implementations in script that are transferable as per https://html.spec.whatwg.org/multipage/#transferable-objects. The implementations are here instead of in script as they need to be passed through the Constellation.

Structs

StructuredSerializedData

A data-holder for serialized data and transferred objects. https://html.spec.whatwg.org/multipage/#structuredserializewithtransfer