servo_constellation_traits::structured_data

Module serializable

This module contains implementations in script that are serializable, as per https://html.spec.whatwg.org/multipage/#serializable-objects. The implementations are here instead of in script as they need to be passed through the Constellation.

Structs

BlobImpl

The data backing a DOM Blob.

BroadcastChannelMsg

Message for communication between the constellation and a global managing broadcast channels.

DomException

A serializable version of the DOMException interface.

DomMatrix

A serializable version of the DOMMatrix/DOMMatrixReadOnly interface.

DomPoint

A serializable version of the DOMPoint/DOMPointReadOnly interface.

DomQuad

A serializable version of the DOMQuad interface.

DomRect

A serializable version of the DOMRect/DOMRectReadOnly interface.

FileBlob

File-based blob

SerializableAesKeyAlgorithm

A serializable version of the AesKeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableAlgorithm

A serializable version of the Algorithm dictionary, used by the SubtleCrypto interface.

SerializableCShakeParams

A serializable version of the CShakeParams dictionary, used by the SubtleCrypto interface.

SerializableCryptoKey

A serializable version of the CryptoKey interface.

SerializableEcKeyAlgorithm

A serializable version of the EcKeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableFile

SerializableFileList

SerializableHmacKeyAlgorithm

A serializable version of the HmacKeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableImageBitmap

A serializable version of the ImageBitmap interface.

SerializableImageData

SerializableIter

An iterator over the variants of Serializable

SerializableKangarooTwelveParams

A serializable version of the KangarooTwelvelParams dictionary, used by the SubtleCrypto interface.

SerializableKeyAlgorithm

A serializable version of the KeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableKmacKeyAlgorithm

A serializable version of the KmacKeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableQuotaExceededError

A serializable version of the QuotaExceededError interface.

SerializableRsaHashedKeyAlgorithm

A serializable version of the RsaHashedKeyAlgorithm dictionary, used the SubtleCrypto interface.

SerializableTurboShakeParams

A serializable version of the TurboShakeParams dictionary, used by the SubtleCrypto interface.

Enums

BlobData

Different backends of Blob

Serializable

All the DOM interfaces that can be serialized.

SerializableCryptoKeyHandle

A serializable version of the Handle type, used by the CryptoKey interface.

SerializableDigestAlgorithm

A serializable version of the DigestAlgorithm type, used the SubtleCrypto interface.

SerializableKeyAlgorithmAndDerivatives

A serializable version of the KeyAlgorithmAndDerivatives type, used by the SubtleCrypto interface.

Traits

BroadcastClone 🔒